[SAFE-232] Move workflow service to the shared library
