+++ projects/back-office/src/app/application/pages/form/form.component.ts
@@ -2,8 +2,8 @@ import { Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { Apollo } from 'apollo-angular';
-import { Subscription, Observable } from 'rxjs';
-import { Form, Page, Step, WhoFormComponent, WhoApplicationService, WhoSnackBarService } from '@who-ems/builder';
+import { Subscription } from 'rxjs';
+import { Form, Page, Step, WhoFormComponent, WhoApplicationService, WhoSnackBarService, WhoWorkflowService } from '@who-ems/builder';
 import {
   GetFormByIdQueryResponse, GET_FORM_BY_ID,
   GetPageByIdQueryResponse, GET_PAGE_BY_ID,
@@ -13,7 +13,6 @@ import {
   EditStepMutationResponse, EDIT_STEP,
   EditPageMutationResponse, EDIT_PAGE
 } from '../../../graphql/mutations';
-import { WorkflowService } from '../../../services/workflow.service';
 
 @Component({
   selector: 'app-form',
@@ -42,7 +41,7 @@ export class FormComponent implements OnInit, OnDestroy {
 
   constructor(
     private applicationService: WhoApplicationService,
-    private workflowService: WorkflowService,
+    private workflowService: WhoWorkflowService,
     private apollo: Apollo,
     private route: ActivatedRoute,
     private router: Router,
